Factors lint list linting into its own module.
This commit also refactors error reporting, so we're not passing
round parameters so much.
